+++ DataFormats/interface/Photon.h
@@ -22,17 +22,17 @@ namespace cat {
     Photon(const reco::LeafCandidate & aPhoton); 
     virtual ~Photon();
 
-    double chargedHadronIso() const { return chargedHadronIso_;  }
-    double puChargedHadronIso() const { return puChargedHadronIso_; }
-    double neutralHadronIso() const { return neutralHadronIso_; }
-    double photonIso() const { return photonIso_; }
+    float chargedHadronIso() const { return chargedHadronIso_;  }
+    float puChargedHadronIso() const { return puChargedHadronIso_; }
+    float neutralHadronIso() const { return neutralHadronIso_; }
+    float photonIso() const { return photonIso_; }
 
   private:
 
-    double chargedHadronIso_;
-    double puChargedHadronIso_;
-    double neutralHadronIso_;
-    double photonIso_;
+    float chargedHadronIso_;
+    float puChargedHadronIso_;
+    float neutralHadronIso_;
+    float photonIso_;
 
   };
 }
